Updated user permission logic to fix issue with wagtail
Now checks multiple groups using an LDAPGroupQuery to verify whether they should be marked as active. This doesn't affect the authorization logic, which still uses an LDAPGroupQuery in local settings.

Should a user be authenticated and wagtail is installed, the user will be added to the moderator and editor group.